requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
Django==3.1.1
django-crontab==0.7.1
djangorestframework==3.11.1
kombu==4.6.11
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
vine==1.3.0
